+++ protobuf/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 6.0.0
 
+* New `GeneratedMessage` extension methods `toTextFormat` and `writeTextFormat`
+  added to convert the message into the [official protocol buffers text
+  format][text format]. ([#1080], [#125])
+
 * **Breaking:** Hide `PbList` and `PbMap` constructors. It is not possible to
   construct these values correctly in user code, so the constructors are now
   private. Existing uses of `PbList` can be replaced by `List` and `PbMap` can
@@ -7,6 +11,9 @@
 
   For immutable lists and maps, you can use `built_value`. ([#1072])
 
+[text format]: https://protobuf.dev/reference/protobuf/textformat-spec/
+[#1080]: https://github.com/google/protobuf.dart/pull/1080
+[#125]: https://github.com/google/protobuf.dart/issues/125
 [#1072]: https://github.com/google/protobuf.dart/pull/1072
 
 ## 5.1.0